// repos/sbs-draft-web-main/utils/types/types.ts
export type PlayerStateInfo = {
    playerId: string
    displayName: string
    team: string
    position: string
    ownerAddress: string
    pickNum: number
    round: number
}

export type PlayerDataProps = {
    playerId: string
    playerStateInfo: PlayerStateInfo
    stats: {
        playerId: string
        averageScore: number
        highestScore: number
        top5Finishes: number
        adp: number
        byeWeek: number
        playersFromTeam: string[]
    }
    ranking: {
        playerId: string
        rank: number
        score: number
    }
}